// src/lib/enums.ts
export enum FormFieldType {
  INPUT = "input",
  PASSWORD_INPUT = "passwordInput",
  TEXTAREA = "textarea",
  CHECKBOX = "checkbox",
  DATE_PICKER = "datePicker",
  DATE_RANGE_PICKER = "dateRangePicker",
  SELECT = "select",
  FILE_INPUT = "fileInput",
  CURRENCY_INPUT = "currencyInput",
  NUMBER_INPUT = "numberInput",
  RADIO_GROUP = "radioGroup",
  SKELETON = "skeleton",
}